+++ EZSwiftExtensionsTests/IntTests.swift
@@ -126,4 +126,12 @@ class IntTests: XCTestCase {
         XCTAssertEqual(UInt.lcm(125, 50), 250)
         XCTAssertEqual(UInt.lcm(65, 52), 260)
     }
+    
+    func testRandomWithinRange() {
+        let closedRange: Range<Int> = 0..<10
+        let randomClosedInt = Int.random(within: closedRange)
+        XCTAssertGreaterThanOrEqual(randomClosedInt, 0)
+        XCTAssertLessThanOrEqual(randomClosedInt, 10)
+    }
+
 }